# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: KRA, KPI & Task Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks
      parameters:
        - name: assignedToId
          in: query
          schema:
            type: integer
        - name: departmentId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: priority
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Not found
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /tasks/{id}/status:
    patch:
      operationId: updateTaskStatus
      tags: [tasks]
      summary: Update task status (manager/HOD/management direct)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskStatusInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}/request-status:
    patch:
      operationId: requestTaskStatus
      tags: [tasks]
      summary: Employee requests a task status change (pending approval)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskStatusRequestInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}/approve-status:
    patch:
      operationId: approveTaskStatus
      tags: [tasks]
      summary: Manager or HOD approves/rejects an employee task status request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskStatusApprovalInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}/comments:
    get:
      operationId: listTaskComments
      tags: [tasks]
      summary: List comments for a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaskComment"
    post:
      operationId: addTaskComment
      tags: [tasks]
      summary: Add a comment to a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskCommentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskComment"

  # ── Dashboard ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Task:
      type: object
      required: [id, title, status, priority, createdById, assignedToId, departmentId, createdAt]
      properties:
        id:               { type: integer }
        title:            { type: string }
        description:      { type: ["string", "null"] }
        status:           { type: string, enum: [pending, in_progress, completed, delayed, approved, rejected] }
        requestedStatus:  { type: ["string", "null"] }
        priority:         { type: string, enum: [high, medium, low] }
        dueDate:          { type: ["string", "null"] }
        completedAt:      { type: ["string", "null"] }
        createdById:      { type: integer }
        createdByName:    { type: string }
        assignedToId:     { type: integer }
        assignedToName:   { type: string }
        departmentId:     { type: integer }
        departmentName:   { type: string }
        isRecurring:      { type: boolean }
        recurringFreq:    { type: ["string", "null"], enum: [daily, weekly, monthly, quarterly, yearly, null] }
        progressPct:      { type: integer }
        createdAt:        { type: string }
    TaskInput:
      type: object
      required: [title, priority, createdById, assignedToId, departmentId]
      properties:
        title:         { type: string, minLength: 1 }
        description:   { type: string }
        priority:      { type: string, enum: [high, medium, low] }
        dueDate:       { type: string }
        createdById:   { type: integer }
        assignedToId:  { type: integer }
        departmentId:  { type: integer }
        isRecurring:   { type: boolean }
        recurringFreq: { type: string, enum: [daily, weekly, monthly, quarterly, yearly] }
    TaskUpdate:
      type: object
      properties:
        title:         { type: string }
        description:   { type: string }
        priority:      { type: string, enum: [high, medium, low] }
        dueDate:       { type: string }
        assignedToId:  { type: integer }
        isRecurring:   { type: boolean }
        recurringFreq: { type: string, enum: [daily, weekly, monthly, quarterly, yearly] }
        progressPct:   { type: integer }
    TaskStatusInput:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [pending, in_progress, completed, delayed, approved, rejected] }
    TaskComment:
      type: object
      required: [id, taskId, authorId, authorName, content, createdAt]
      properties:
        id:         { type: integer }
        taskId:     { type: integer }
        authorId:   { type: integer }
        authorName: { type: string }
        content:    { type: string }
        createdAt:  { type: string }
    TaskCommentInput:
      type: object
      required: [authorId, content]
      properties:
        authorId: { type: integer }
        content:  { type: string, minLength: 1 }

    # ── Dashboard ────────────────────────────────────────────────────────────────
...
    TaskStatusRequestInput:
      type: object
      required: [status]
      properties:
        status:      { type: string, enum: [in_progress, completed, delayed] }
        progressPct: { type: integer }
    TaskStatusApprovalInput:
      type: object
      required: [approved]
      properties:
        approved: { type: boolean }
        reason:   { type: string }
...